refactor : mapping table 구현에 따른 코드 변경.
세부 사항
- 모호한 변수 네이밍 수정.
