+++ src/main/java/controller/services/BridgeNumberGenerator.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.services;
 
 @FunctionalInterface
 public interface BridgeNumberGenerator {
