+++ src/main/java/controller/services/BridgeRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.services;
 
 import camp.nextstep.edu.missionutils.Randoms;
